backend/tg-channel-import/requirements.txt
psycopg2-binary
boto3
telethon
